MFA work aborted because of known jython issue SAXParser ClassNotFound. Implemented a rudimentary universalIPRotator.py which require some manual AWS operations prior to using it
